ftdi_eeprom: support channel configuration
Previously, the channels could not be configured and were
hard set to type UART and driver VCP and further to no RS485
functionality on chips with support for this feature.

With the new 'chX_*' config file options the ftdi_eeprom tool
is now abel to change the channel type, set the driver autoload
decision (VCP od D2XX) and enable or disable RS485.

The new config file options are:

 * ch[a,b]_type      - string of: UART, FIFO, OPTO, CPU, FT1284
 * ch[a,b,c,d]_vcp   - bool: true for VCP, false for D2XX
 * ch[a,b,c,d]_rs485 - bool: true for RS485 enabled

Signed-off-by: Stephan Linz <linz@li-pro.net>